# static/swagger.yaml
openapi: 3.0.3
info:
  title: E-commerce Recommendation System API
  description: |
    This API provides AI-powered recommendation and search capabilities for e-commerce platforms.
    
    Features include:
    * Content-based product recommendations
    * Personalized user recommendations
    * NLP-based semantic product search
    * Image-based product search
    * Multimodal search (text + image)
    * A/B testing framework
    * User activity tracking
  version: 1.0.0
  contact:
    name: Aayush Sharma
    email: aayush.75way@gmail.com
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.yourdomain.com/api
    description: Production server
...
paths:
  /info:
    get:
      tags:
        - info
      summary: Get API information
      description: Returns information about the API and available components
      operationId: getApiInfo
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiInfo'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /recommendations/content:
    get:
      tags:
        - recommendations
      summary: Get content-based recommendations
      description: Get recommendations for similar products based on a given product
      operationId: getContentRecommendations
      parameters:
        - name: item_name
          in: query
          description: Name of the product to get recommendations for
          required: true
          schema:
            type: string
        - name: top_n
          in: query
          description: Number of recommendations to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentRecommendationResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /recommendations/user:
    get:
      tags:
        - recommendations
      summary: Get personalized recommendations
      description: Get personalized recommendations for a user, optionally based on a specific product
      operationId: getUserRecommendations
      parameters:
        - name: user_id
          in: query
          description: User ID to get recommendations for
          required: true
          schema:
            type: integer
            format: int64
        - name: item_name
          in: query
          description: Name of the product to use for content-based filtering (optional)
          required: false
          schema:
            type: string
        - name: top_n
          in: query
          description: Number of recommendations to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecommendationResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /search/nlp:
    get:
      tags:
        - search
      summary: Natural language search
      description: Search products using natural language processing
      operationId: nlpSearch
      parameters:
        - name: query
          in: query
          description: Natural language search query
          required: true
          schema:
            type: string
        - name: top_n
          in: query
          description: Number of results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: user_id
          in: query
          description: User ID for personalized results (optional)
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NLPSearchResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /search/image:
    post:
      tags:
        - search
      summary: Image-based search
      description: Search products using image similarity
      operationId: imageSearch
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file for similarity search
                top_n:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 10
                  description: Number of results to return
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageSearchResponse'
        '400':
          description: Invalid parameters or image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /search/multimodal:
    post:
      tags:
        - search
      summary: Multimodal search
      description: Search products using both image and text
      operationId: multimodalSearch
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
                - text_query
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file for similarity search
                text_query:
                  type: string
                  description: Text search query
                top_n:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 10
                  description: Number of results to return
                image_weight:
                  type: number
                  minimum: 0
                  maximum: 100
                  default: 50
                  description: Weight for image results (0-100)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultimodalSearchResponse'
        '400':
          description: Invalid parameters or image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /trending:
    get:
      tags:
        - trending
      summary: Get trending products
      description: Get a list of trending products
      operationId: getTrendingProducts
      parameters:
        - name: top_n
          in: query
          description: Number of products to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendingProductsResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /activity:
    post:
      tags:
        - activity
      summary: Record user activity
      description: Record user activity for improved recommendations
      operationId: recordUserActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserActivity'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /ab_testing/assign:
    get:
      tags:
        - ab-testing
      summary: Assign user to A/B test
      description: Assign a user to A/B test variants and get recommendation weights
      operationId: assignABTest
      parameters:
        - name: user_id
          in: query
          description: User ID to assign
          required: true
          schema:
            type: integer
            format: int64
        - name: test_id
          in: query
          description: Specific test ID to assign to (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ABTestingAssignResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
  /products:
    get:
      tags:
        - products
      summary: Get products
      description: Get products with filtering options
      operationId: getProducts
      parameters:
        - name: category
          in: query
          description: Filter by category
          required: false
          schema:
            type: string
        - name: brand
          in: query
          description: Filter by brand
          required: false
          schema:
            type: string
        - name: query
          in: query
          description: Search query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Number of products to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Pagination offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Enter your API key with the format: "Bearer YOUR_API_KEY"'
  schemas:
...
    ApiInfo:
      type: object
      properties:
        status:
          type: string
          example: "success"
        service:
          type: string
          example: "E-commerce Recommendation API"
        model_info:
          type: object
          properties:
            product_count:
              type: integer
              example: 4090
            components_available:
              type: object
              properties:
                recommendation_system:
                  type: boolean
                  example: true
                nlp_search:
                  type: boolean
                  example: true
                image_search:
                  type: boolean
                  example: true
                multimodal_search:
                  type: boolean
                  example: true
                ab_testing:
                  type: boolean
                  example: true
            version:
              type: string
              example: "1.0.0"
    ContentRecommendationResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        item:
          type: string
          example: "Smartphone X"
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    UserRecommendationResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        user_id:
          type: integer
          example: 123
        item:
          type: string
          example: "Smartphone X"
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/ProductWithRecommendationType'
...
    NLPSearchResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        query:
          type: string
          example: "affordable smartphones with good camera under $300"
        detected_attributes:
          $ref: '#/components/schemas/QueryAttributes'
        results:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    ImageSearchResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        image:
          type: string
          example: "uploaded_image.jpg"
        results:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...
    TrendingProductsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        trending_products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Error message description"
        traceback:
          type: string
          example: "Detailed error traceback (only in development mode)"